chore: move selector-utils into selectors directory
