+++ src/state/slices/tradeQuoteSlice/selectors.ts
@@ -19,11 +19,11 @@ import { TradeQuoteRequestError, TradeQuoteWarning } from 'state/apis/swapper'
 import { validateQuoteRequest } from 'state/apis/swapper/helpers/validateQuoteRequest'
 import { selectIsTradeQuoteApiQueryPending } from 'state/apis/swapper/selectors'
 import type { ReduxState } from 'state/reducer'
-import { createDeepEqualOutputSelector } from 'state/selector-utils'
 import {
   selectIsWalletConnected,
   selectWalletSupportedChainIds,
 } from 'state/selectors/common-selectors'
+import { createDeepEqualOutputSelector } from 'state/selectors/selector-utils'
 import { selectFeeAssetById } from 'state/slices/assetsSlice/selectors'
 import {
   selectFirstHopSellAccountId,
